refactor(policy): move evaluation engine to its owner (#498 W1-W4)
W1 — evaluation engine moved to packages/policy:
- Policy.evaluate + PolicyDecision.fromEvaluation moved out of protocol into
  packages/policy/src/permission-evaluate.ts as evaluatePermission /
  decisionFromEvaluation (same signatures/behavior). Protocol keeps the zod
  vocabulary and the ReDoS-safety chain; isSafeInputPattern and
  MAX_INPUT_LENGTH are exported from PolicyPermission (no duplication) and
  the PolicyPermission namespace now rides the policy barrel.
- Rewired production callers: 4, not the inventoried 2 — gateway stage 0
  (#718) added apps/server/src/channel/authn/decision.ts and
  apps/server/src/tool/mcp/mcp-prefix-guard.ts since the inventory head, so
  they are rewired alongside ingress-authority and tool-permission-policy.
- Evaluate-behavior tests moved to packages/policy/test/
  permission-evaluate.test.ts (+ decisionFromEvaluation mapping pins);
  schema-only tests stay in protocol; the module-surface pin update is the
  zero-old-family receipt.
- lint-guards canonical-evaluator rule now points at the new leaf and
  requires evaluatePermission(); fixed the stale
  ingress-authority-evaluation.ts entry and added tool-permission-policy.

W2 — agent facade collapsed:
- Deleted the do-nothing PolicyEngine.create wrapper in
  agent/core/policy/index.ts. PolicyEngine there is now a typed const
  binding of the same generic engine object from @openomni/policy (a type
  application, not a wrapper function or a new object), so the ~40 existing
  value consumers keep compiling against the generic engine directly. The
  PolicyEngineInstance alias collapsed to
  PolicyEngineInstanceGeneric<PolicyContext>; the Omit hack existed only
  for the wrapper.

W3 — RuntimeResource folded into Policy.Resource:
- Namespace renamed to PolicyResource, exposed as Policy.Resource via
  export import; the sibling export is gone. Zero field/shape changes
  (descriptor rides bus events; wire-frozen). 29 consumer files rewired;
  zero RuntimeResource imports/exports remain (rg-verified).

W4 — dead vocabulary removed (proofs in #498):
- PolicyPoint.MigrationMapping (test-only) deleted; migration-mapping.test
  deleted and point-registry/point-audit/point-dispatch/dispatch-schema/
  module-surface tests pruned.
- Policy.Definition.timing removed; both .omit({ timing: true }) call sites
  simplified (policy registration-validation, openomni policy-registration);
  channel-authn definitions no longer stamp it. Policy.Timing alias and its
  engine/agent consumers untouched.
- Dispatch.ActorContext.permissions and its point-input-schemas
  dispatchActor clone removed in lockstep; input-schema-parity fixture
  fixed. Ingress.AgentDef/ExecutionRequest wire permissions untouched.

Schema snapshot regenerated via lint-tools --update (the sign-off surface):
RuntimeResource.* -> Policy.Resource.*, Policy.Definition -timing,
Dispatch.ActorContext -permissions; the regen also records previously
unsnapshotted additive types (Gateway.*, LedgerAppend.*, Ipc.*) and
normalizes serializer formatting.

Verification: build green, check-types 14/14, lint-tools clean after
update, lint-guards + check-deps clean, protocol 610 pass, policy 110
pass, agent 416 pass, openomni ingress/dispatch/messaging 331 pass and
execution-runtime/policy/tool 345 pass, apps/server channel-authn + mcp
29 pass, #498 verification suite 22 pass.
