+++ packages/agent/src/runtime/mcp/client-descriptor.ts
@@ -1,11 +1,11 @@
-import type { RuntimeResource } from "@openomni/protocol";
+import type { Policy } from "@openomni/protocol";
 import { Tool } from "@openomni/protocol";
 
 type McpToolSpec = Tool.Spec & {
-  readonly descriptor: RuntimeResource.Descriptor;
+  readonly descriptor: Policy.Resource.Descriptor;
 };
 
-function createMcpToolDescriptor(serverId: string, remoteName: string): RuntimeResource.Descriptor {
+function createMcpToolDescriptor(serverId: string, remoteName: string): Policy.Resource.Descriptor {
   return {
     id: `tool:mcp:${serverId}:${remoteName}`,
     kind: "tool",
