+++ packages/openomni/src/execution-runtime/tool/types.ts
@@ -1,4 +1,4 @@
-import type { Policy, RuntimeResource, Tool, ToolSelection } from "@openomni/protocol";
+import type { Policy, Tool, ToolSelection } from "@openomni/protocol";
 
 export type ToolCategory = "system" | "agent" | "mcp";
 export type ToolMetaValue = boolean | ((input: unknown) => boolean);
@@ -27,7 +27,7 @@ export interface NativeTool {
   isDestructive: ToolMetaValue;
   isConcurrencySafe: ToolMetaValue;
   labels?: readonly string[];
-  descriptor?: RuntimeResource.Descriptor;
+  descriptor?: Policy.Resource.Descriptor;
   source?: ToolSource;
   category?: ToolSelection.Category;
   implicitInputs?: Readonly<Record<string, ImplicitInputSource>>;
