+++ packages/policy/test/point-audit.test.ts
@@ -6,8 +6,6 @@ describe("PolicyEngine canonical point audit", () => {
   test("stamps evaluated and composed events with the dispatched point", async () => {
     // Given
     const pointId = "dispatch.action.pre" as const;
-    const timing = Policy.Timing.DISPATCH_AUTHORIZE;
-    const originalMapping = Policy.PolicyPoint.MigrationMapping[timing];
     const traceContext = {
       traceId: "trace-canonical-audit",
       sessionId: "session-canonical-audit",
@@ -27,20 +25,15 @@ describe("PolicyEngine canonical point audit", () => {
       fn: () => PolicyDecision.allow({ policyId: "canonical-audit" }),
     });
 
-    Reflect.set(Policy.PolicyPoint.MigrationMapping, timing, ["run.lifecycle.pre"]);
-    try {
-      // When
-      await engine.dispatchPoint(pointId, {
-        actor: { kind: "system", actorId: "system:test" },
-        dispatchId: "dispatch-audit",
-        action: "resident.ask",
-        target: { kind: "resident" },
-        sessionId: traceContext.sessionId,
-        runId: traceContext.runId,
-      });
-    } finally {
-      Reflect.set(Policy.PolicyPoint.MigrationMapping, timing, originalMapping);
-    }
+    // When
+    await engine.dispatchPoint(pointId, {
+      actor: { kind: "system", actorId: "system:test" },
+      dispatchId: "dispatch-audit",
+      action: "resident.ask",
+      target: { kind: "resident" },
+      sessionId: traceContext.sessionId,
+      runId: traceContext.runId,
+    });
 
     // Then
     const evaluated = PolicyEvent.Evaluated.schema.parse(
