+++ packages/protocol/src/policy/permission.ts
@@ -2,8 +2,8 @@ import { z } from "zod";
 
 export namespace PolicyPermission {
   const MAX_REGEX_PATTERN_LENGTH = 200;
-  const MAX_INPUT_LENGTH = 10_000;
-  const POLICY_ID = "guardrail.permission";
+  /** Evaluation-side input cap, shared with the evaluator in @openomni/policy. */
+  export const MAX_INPUT_LENGTH = 10_000;
 
   // Label source enumeration: where a label originates from
   // Labels use source.category naming convention to prevent namespace collisions
@@ -20,7 +20,6 @@ export namespace PolicyPermission {
   export type LabelEntry = z.infer<typeof LabelEntry>;
 
   const PermissionDecision = z.enum(["allow", "deny", "require_approval"]);
-  type PermissionDecision = z.infer<typeof PermissionDecision>;
 
   export const InputRule = z.object({
     toolPattern: z.string(),
@@ -124,125 +123,16 @@ export namespace PolicyPermission {
     );
   }
 
-  function isSafeInputPattern(pattern: string): boolean {
+  /**
+   * ReDoS-safety predicate for `InputRule.pattern`. Owned here because the
+   * zod refine above validates at the schema boundary; the evaluation engine
+   * (`evaluatePermission` in @openomni/policy) re-checks it at runtime.
+   */
+  export function isSafeInputPattern(pattern: string): boolean {
     return (
       pattern.length <= MAX_REGEX_PATTERN_LENGTH &&
       isRegexSyntaxValid(pattern) &&
       !hasUnsafeQuantifier(pattern)
     );
-  }
-
-  function matchesPattern(resource: string, pattern: string): boolean {
-    if (pattern === "*") return true;
-    if (pattern.endsWith(".*")) return resource.startsWith(`${pattern.slice(0, -2)}.`);
-    return resource === pattern;
-  }
-
-  function findMatchingLabel(
-    labels: readonly string[] | undefined,
-    patterns: readonly string[] | undefined,
-  ): string | undefined {
-    if (!patterns || patterns.length === 0) return undefined;
-    for (const pattern of patterns) {
-      if (labels?.some((label) => matchesPattern(label, pattern))) return pattern;
-    }
-    return undefined;
-  }
-
-  type InputMatchResult = "match" | "miss" | "unsafe";
-
-  function matchesInputField(
-    input: Record<string, unknown> | undefined,
-    field: string,
-    pattern: string,
-  ): InputMatchResult {
-    if (!isSafeInputPattern(pattern)) return "unsafe";
-
-    const raw = String(input?.[field] ?? "");
-    const value = raw.length > MAX_INPUT_LENGTH ? raw.slice(0, MAX_INPUT_LENGTH) : raw;
-
-    return new RegExp(pattern).test(value) ? "match" : "miss";
-  }
-
-  function verdict(
-    decision: PermissionDecision,
-    reason: string,
-    matchedPattern?: string,
-  ): EvaluationResult {
-    const action: EvaluationResult["action"] = decision === "allow" ? "continue" : "abort";
-    return matchedPattern === undefined
-      ? { action, decision, reason, policyId: POLICY_ID }
-      : { action, decision, reason, policyId: POLICY_ID, matchedPattern };
-  }
-
-  export function evaluate(
-    permission: Permission | undefined,
-    request: EvaluationRequest,
-  ): EvaluationResult {
-    if (!permission) return verdict("allow", "default_allow");
-    if (permission.action !== request.action) return verdict("deny", "action_mismatch");
-
-    const inputRules = [...(permission.inputRules ?? [])].sort(
-      (a, b) => (b.priority ?? 0) - (a.priority ?? 0),
-    );
-
-    for (const rule of inputRules) {
-      if (!matchesPattern(request.resource, rule.toolPattern)) continue;
-
-      const inputMatch = matchesInputField(request.input, rule.field, rule.pattern);
-      if (inputMatch === "unsafe") {
-        return verdict("deny", "unsafe_input_rule", rule.toolPattern);
-      }
-      if (inputMatch === "match") {
-        return verdict(rule.action, rule.reason ?? `input_rule_${rule.action}`, rule.toolPattern);
-      }
-    }
-
-    const deniedBy = permission.denylist?.find((pattern) =>
-      matchesPattern(request.resource, pattern),
-    );
-    if (deniedBy) return verdict("deny", "denylist", deniedBy);
-
-    const deniedByLabel = findMatchingLabel(request.resourceLabels, permission.denyLabels);
-    if (deniedByLabel) return verdict("deny", "deny_label", deniedByLabel);
-
-    const requiresApprovalBy = permission.requireApproval?.find((pattern) =>
-      matchesPattern(request.resource, pattern),
-    );
-    if (requiresApprovalBy) {
-      return verdict("require_approval", "require_approval", requiresApprovalBy);
-    }
-
-    const requiresApprovalByLabel = findMatchingLabel(
-      request.resourceLabels,
-      permission.requireApprovalLabels,
-    );
-    if (requiresApprovalByLabel) {
-      return verdict("require_approval", "require_approval_label", requiresApprovalByLabel);
-    }
-
-    if (permission.allowlist !== undefined) {
-      const allowedBy = permission.allowlist.find((pattern) =>
-        matchesPattern(request.resource, pattern),
-      );
-
-      if (allowedBy) return verdict("allow", "allowlist", allowedBy);
-
-      return verdict(
-        "deny",
-        permission.allowlist.length === 0 ? "allowlist_empty" : "allowlist_miss",
-      );
-    }
-
-    if (permission.allowLabels !== undefined) {
-      const allowedByLabel = findMatchingLabel(request.resourceLabels, permission.allowLabels);
-      if (allowedByLabel) return verdict("allow", "allow_label", allowedByLabel);
-      return verdict(
-        "deny",
-        permission.allowLabels.length === 0 ? "allow_labels_empty" : "allow_labels_miss",
-      );
-    }
-
-    return verdict("allow", "default_allow");
   }
 }
